+++ preboot-query/src/test/java/io/preboot/query/testdata/TestEventRepository.java
@@ -1,0 +1,5 @@
+package io.preboot.query.testdata;
+
+import io.preboot.query.FilterableRepository;
+
+public interface TestEventRepository extends FilterableRepository<TestEvent, Long> {}