+++ preboot-query/src/test/java/io/preboot/query/testdata/TestOrderWithEnumRepositoryImpl.java
@@ -1,0 +1,12 @@
+package io.preboot.query.testdata;
+
+import io.preboot.query.FilterableFragmentContext;
+import io.preboot.query.FilterableFragmentImpl;
+import org.springframework.stereotype.Repository;
+
+@Repository
+class TestOrderWithEnumRepositoryImpl extends FilterableFragmentImpl<TestOrderWithEnum, Long> {
+    public TestOrderWithEnumRepositoryImpl(FilterableFragmentContext context) {
+        super(context, TestOrderWithEnum.class);
+    }
+}